Move tests of node in its dedicated file
